geosense/requirements.txt
# Field Inspection System - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
streamlit>=1.28.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=10.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0

# Geolocation & Mapping
folium>=0.14.0
streamlit-folium>=0.15.0
geopy>=2.3.0

# Utilities
python-dateutil>=2.8.0